Remove metabase/lib/schema dependency in metabase-lib (#25905)
