+++ frontend/src/metabase/lib/schema_metadata.js
@@ -1,16 +1,4 @@
 import { field_semantic_types_map } from "metabase/lib/core";
-import { getFieldType } from "metabase-lib/lib/types/utils/isa";
-import {
-  TEMPORAL,
-  LOCATION,
-  COORDINATE,
-  FOREIGN_KEY,
-  PRIMARY_KEY,
-  STRING,
-  STRING_LIKE,
-  NUMBER,
-  BOOLEAN,
-} from "metabase-lib/lib/types/constants";
 
 export function foreignKeyCountsByOriginTable(fks) {
   if (fks === null || !Array.isArray(fks)) {
@@ -30,22 +18,6 @@ export function foreignKeyCountsByOriginTable(fks) {
 
       return prev;
     }, {});
-}
-
-export const ICON_MAPPING = {
-  [TEMPORAL]: "calendar",
-  [LOCATION]: "location",
-  [COORDINATE]: "location",
-  [STRING]: "string",
-  [STRING_LIKE]: "string",
-  [NUMBER]: "int",
-  [BOOLEAN]: "io",
-  [FOREIGN_KEY]: "connections",
-  [PRIMARY_KEY]: "label",
-};
-
-export function getIconForField(field) {
-  return ICON_MAPPING[getFieldType(field)] || "unknown";
 }
 
 export function getSemanticTypeIcon(semanticType, fallback) {
